# api/docs/testing/backend-mvp-a.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Backend MVP-A Course/Lesson visibility - TestSprite handoff",
    "version": "1.0.0",
    "description": "Read-only generation contract. Real Student/Admin credentials are separate. Missing producers block dependents. This file is not evidence of a TestSprite run."
  },
  "servers": [
    {
      "url": "http://72.61.87.212:8080/api",
      "description": "Owner-supplied server; verify deployment and target before authenticated acceptance."
    }
  ],
  "paths": {
...
    "/v1/admin/students": {
      "get": {
        "tags": [
          "Students"
        ],
        "parameters": [
          {
            "name": "brand",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "medway",
                "elite",
                "nexus"
              ]
            },
            "description": "Commercial brand scope only; BUC and Delta are academic institutions, not brands."
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 120
            },
            "description": "Matches opaque student-profile ID, or the private name, email, or student code without returning raw email/code."
          },
          {
            "name": "academicLevel",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Academic level ID from the private academic placement source."
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "pending",
                "disabled",
                "suspended"
              ]
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "summary": "List read-only student profile records across commercial brands",
        "description": "Requires a verified admin identity and persisted administrative permission. A student bearer does not grant admin access. Missing bearer is rejected before route/query validation; a syntactically supplied bearer with an invalid route/query can return 400 before identity resolution.",
        "security": [
          {
            "AdminBearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Read-only student list with private fields minimized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminStudentsListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid route or query",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found in authorized scope",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "503": {
            "description": "Read source unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "x-mvp-auth": "admin",
        "x-mvp-producer": [
          "admin_student_id"
        ],
        "x-mvp-bindings": {},
        "x-mvp-selection": "Select only the preapproved controlled student from data.data[]; retain id privately. Empty lists are valid.",
        "x-mvp-missing-dependency": "blocked: do not send a request or invent an identifier"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminStudentsListResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "const": true
          },
          "correlationId": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "data": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AdminStudentReadItem"
                }
              },
              "pagination": {
                "type": "object"
              }
            }
          }
        }
      },
      "AdminStudentReadItem": {
        "type": "object",
        "additionalProperties": false,
        "description": "Authorized administrative student read. Full name is private admin data; email and student code are masked. Phone, device, session, activity, and access facts remain null or empty until an approved source exists.",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "platform": {
            "type": "object"
          },
          "displayName": {
            "type": [
              "string",
              "null"
            ]
          },
          "emailMasked": {
            "type": [
              "string",
              "null"
            ]
          },
          "studentIdMasked": {
            "type": [
              "string",
              "null"
            ]
          },
          "phoneMasked": {
            "type": [
              "null"
            ]
          },
          "academicInstitution": {
            "type": [
              "string",
              "null"
            ],
            "description": "Academic institution such as BUC or Delta; never a commercial brand."
          },
          "academicLevel": {
            "type": [
              "string",
              "null"
            ]
          },
          "academicLevelId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "academicSemester": {
            "type": [
              "string",
              "null"
            ]
          },
          "program": {
            "type": [
              "string",
              "null"
            ]
          },
          "expectedGraduationDate": {
            "type": [
              "string",
              "null"
            ],
            "format": "date"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "disabled"
            ]
          },
          "activeDeviceCount": {
            "type": [
              "null"
            ]
          },
          "activeSessionCount": {
            "type": [
              "null"
            ]
          },
          "lastSeenAt": {
            "type": [
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "AdminBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Real admin access token with persisted admin permission."
      },
      "StudentBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Real student access token with active student membership and placement. An identity with an admin profile is rejected."
      }
    }
  }
}